+++ inst/examples/get_mallows_loglik_example.R
@@ -8,15 +8,16 @@ mydata <- sample_mallows(
   metric = "kendall")
 
 # Compute the likelihood and log-likelihood values under the true model...
-lik_db_mix(
+get_mallows_loglik(
   rho = rbind(1:n_items, 1:n_items),
   alpha = c(10, 10),
   weights = c(0.5, 0.5),
   metric = "kendall",
-  rankings = mydata
+  rankings = mydata,
+  log = FALSE
   )
 
-lik_db_mix(
+get_mallows_loglik(
   rho = rbind(1:n_items, 1:n_items),
   alpha = c(10, 10),
   weights = c(0.5, 0.5),
@@ -27,16 +28,17 @@ lik_db_mix(
 
 # or equivalently, by using the frequency distribution
 freq_distr <- rank_freq_distr(mydata)
-lik_db_mix(
+get_mallows_loglik(
   rho = rbind(1:n_items, 1:n_items),
   alpha = c(10, 10),
   weights = c(0.5, 0.5),
   metric = "kendall",
   rankings = freq_distr[, 1:n_items],
-  obs_freq = freq_distr[, n_items + 1]
+  obs_freq = freq_distr[, n_items + 1],
+  log = FALSE
   )
 
-lik_db_mix(
+get_mallows_loglik(
   rho = rbind(1:n_items, 1:n_items),
   alpha = c(10, 10),
   weights = c(0.5, 0.5),