+++ NEWS.md
@@ -2,9 +2,12 @@
 
 # Development versions
 
-* Fixes a bug which caused assess_convergence() to fail with 'parameter = "cluster_probs"'.
+* Fixes a bug which caused assess_convergence() to fail with 'parameter = 
+  "cluster_probs"'.
 * BayesMallows no longer depends on 'dplyr'.
 * Quite extensive internal refactoring of C++ code.
+* Function lik_db_mix has been renamed to get_mallows_loglik. lik_db_mix still
+  exists as deprecated.
 
 # BayesMallows 1.1.2